captcha-recognition/requirements.txt
h5py==2.7.1
numpy==1.14.2
scipy==1.0.0
protobuf==3.5.2
Pillow==5.0.0
tensorflow==1.6.0
tensorboard==1.6.0
Keras==2.1.5
